pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "numpy>=2.2.2",
    "openai>=1.61.0",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "pyactiveresource>=2.2.2",
    "shopifyapi>=12.7.0",
    "sqlalchemy>=2.0.37",
    "streamlit>=1.42.0",
    "twilio>=9.4.4",
]
